Fix Service Stop
